put new command into php.bee.inc
